+++ packages/client/ui/Button/Button.tsx
@@ -41,7 +41,7 @@ export interface ButtonProps extends React.ButtonHTMLAttributes<HTMLButtonElemen
   shape?: Shape
 }
 
-const Button = React.forwardRef<HTMLButtonElement, ButtonProps>(
+export const Button = React.forwardRef<HTMLButtonElement, ButtonProps>(
   ({className, variant, size = 'default', shape = 'default', asChild = false, ...props}, ref) => {
     const Comp = asChild ? Slot : 'button'
     return (
@@ -63,5 +63,3 @@ const Button = React.forwardRef<HTMLButtonElement, ButtonProps>(
 )
 
 Button.displayName = 'Button'
-
-export {Button}