+++ packages/server/graphql/private/mutations/stripeFailPayment.ts
@@ -101,7 +101,6 @@ const stripeFailPayment: MutationResolvers['stripeFailPayment'] = async (
   )
 
   await r({
-    update: r.table('Invoice').get(invoiceId).update({status: 'FAILED'}),
     insert: r.table('Notification').insert(notifications)
   }).run()